Mac sergey

See merge request a.sergey/beatcovid!38

+++ BeatCovid/include/Entities/Transports/SafeTransport.h
@@ -14,7 +14,7 @@ class SafeTransport : public Transport {
  public:
   explicit SafeTransport(float x, float y, int height, int weight, const std::string &name, std::vector<float> config);
   void Update(float time, std::vector<Object> &obj) override;
-  float GetFuel() const;
+  int GetFuel() const;
 
  private:
   float fuel;
